requirements.txt
annotated-types
anyio
click
exceptiongroup
fastapi
h11
idna
numpy
pandas 
pydantic
pydantic-core
python-dateutil
pytz
scikit-learn  
scikit-surprise
scipy
six
sniffio
starlette
typing-extensions
tzdata
uvicorn
pyarrow

